fixed a bug where write()'s offset param was being passed into C as a long rather than a ulong. This allowed the caller to do stupid things like passing a negative offset.
THe fixed code uses the 'k' format specifier which matches the format specifier
for offset in the read() method. offset has to be a ulong in order to address
really big memory segments. It makes the Python code weaker, though,
because the offset passed is converted to a ulong without overflow checking.