revert change to mq.c (aab5846e2e1813581c30c3b3a7f7a5eab326cc24)
I thought that commit was just a whitespace change, but when I created a PR
to merge this change, GitHub shows the lines as being entirely new.

This file hasn't changed in five years. My guess is that I added the lines
(a DPRINTF statement) while experimenting sometime in that period and never
bothered to commit them. Since they're for debugging only, I'm just going
to get rid of them. I can add them back if I need them.