Remove travis support
Travis hasn't worked in a long time, I'm not sure why. It will
soon be replaced in this project by cibuildwheel (see #26) so
I'm just eliminating travis rather than trying to fix support for it.

This also gets rid of setup.cfg which I should have deleted in PR #32.
setup.cfg referenced flake8 (which I'm not currently using), and that
kind of config should be in pyproject.toml now