rename constant SEMAPHORE_VALUE_MAX to SEM_VALUE_MAX
There's a subtle importance to this, because SEM_VALUE_MAX is already
defined in some system header files, whereas SEMAPHORE_VALUE_MAX never
was. That change is deliberate. If it's already #defined by the system,
sysv_ipc will use that value. Otherwise it will use whatever is in
system_info.h.

In addition, I removed one superfluous check from my code. It was
there to prevent the semaphore's value from going out of range, but
of course the operating system's own semaphore implementation
prevents that. I removed the check from my code and now the error
is handled in `sem_set_error()`